Refactored FC components
